+++ spharpy/__init__.py
@@ -8,9 +8,9 @@ __version__ = '0.6.2'
 
 from .classes.sh import SphericalHarmonics
 from .classes.coordinates import SamplingSphere
-from .classes.audio import SphericalHarmonicSignal
-from .classes.audio import SphericalHarmonicTimeData
-from .classes.audio import SphericalHarmonicFrequencyData
+from .classes.sh_audio import SphericalHarmonicSignal
+from .classes.sh_audio import SphericalHarmonicTimeData
+from .classes.sh_audio import SphericalHarmonicFrequencyData
 from . import spherical
 from . import samplings
 from . import plot
